CÓDIGO NÃO ESTÁ FUNCIONANDO, fiz a alteração de int para long das variáveis, mandei uma possível correção de erro no zap

+++ src/main/java/com/grupo2/Video.java
@@ -1,22 +1,22 @@
 package com.grupo2;
 
 public class Video {
-    private int numColuna;
+    private long numColuna;
     private String claim_Status;
-    private int video_id;
-    private int video_duration_sec;
+    private long video_id;
+    private long video_duration_sec;
     private String video_transcription_text;
     private String verified_status;
     private String author_ban_status;
-    private int video_view_count;
-    private int video_like_count;
-    private int video_share_count;
-    private int video_download_count;
+    private long video_view_count;
+    private long video_like_count;
+    private long video_share_count;
+    private long video_download_count;
 
     
-    public Video(int numColuna, String claim_Status, int video_id, int video_duration_sec, String video_transcription_text,
-            String verified_status, String author_ban_status, int video_view_count, int video_like_count,
-            int video_share_count, int video_download_count, int video_comment_count) {
+    public Video(long numColuna, String claim_Status, long video_id, long video_duration_sec, String video_transcription_text,
+            String verified_status, String author_ban_status, long video_view_count, long video_like_count,
+            long video_share_count, long video_download_count, long video_comment_count) {
         this.numColuna = numColuna;
         this.claim_Status = claim_Status;
         this.video_id = video_id;
@@ -39,19 +39,19 @@ public class Video {
         this.claim_Status = claim_Status;
     }
 
-    public int getVideo_id() {
+    public long getVideo_id() {
         return video_id;
     }
 
-    public void setVideo_id(int video_id) {
+    public void setVideo_id(long video_id) {
         this.video_id = video_id;
     }
 
-    public int getVideo_duration_sec() {
+    public long getVideo_duration_sec() {
         return video_duration_sec;
     }
 
-    public void setVideo_duration_sec(int video_duration_sec) {
+    public void setVideo_duration_sec(long video_duration_sec) {
         this.video_duration_sec = video_duration_sec;
     }
 
@@ -79,54 +79,54 @@ public class Video {
         this.author_ban_status = author_ban_status;
     }
 
-    public int getVideo_view_count() {
+    public long getVideo_view_count() {
         return video_view_count;
     }
 
-    public void setVideo_view_count(int video_view_count) {
+    public void setVideo_view_count(long video_view_count) {
         this.video_view_count = video_view_count;
     }
 
-    public int getVideo_like_count() {
+    public long getVideo_like_count() {
         return video_like_count;
     }
 
-    public void setVideo_like_count(int video_like_count) {
+    public void setVideo_like_count(long video_like_count) {
         this.video_like_count = video_like_count;
     }
 
-    public int getVideo_share_count() {
+    public long getVideo_share_count() {
         return video_share_count;
     }
 
-    public void setVideo_share_count(int video_share_count) {
+    public void setVideo_share_count(long video_share_count) {
         this.video_share_count = video_share_count;
     }
 
-    public int getVideo_download_count() {
+    public long getVideo_download_count() {
         return video_download_count;
     }
 
-    public void setVideo_download_count(int video_download_count) {
+    public void setVideo_download_count(long video_download_count) {
         this.video_download_count = video_download_count;
     }
 
-    public int getVideo_comment_count() {
+    public long getVideo_comment_count() {
         return video_comment_count;
     }
 
-    public void setVideo_comment_count(int video_comment_count) {
+    public void setVideo_comment_count(long video_comment_count) {
         this.video_comment_count = video_comment_count;
     }
 
-    private int video_comment_count;
+    private long video_comment_count;
 
 
-    public int getNumColuna() {
+    public long getNumColuna() {
         return numColuna;
     }
 
-    public void setNumColuna(int numColuna) {
+    public void setNumColuna(long numColuna) {
         this.numColuna = numColuna;
     }
